Fixes deploy script.
Major fixes in fabfile.
Adds deploy cmd
Adds domains to ALLOWED_HOSTS in default.py
Adds new django1.7 uwsgi loader to prod.wsgi

+++ fabfile.py
@@ -7,24 +7,32 @@ from fabric.api import local, cd, sudo
 from fabric.contrib.files import exists
 
 from fabconfig import env
+from fabconfig import prod
 
-# def deploy should do the following:
-# get branch
-# fetch origin master
-# merge master locally
-# push to origin
-# create the gir archive
-# push the archive
-# unzip the archive on the machine
-# create the virtualenv
-# static files
-# deploy nginx
-# restart nginx
-# reload supervisor
-# empty /tmo/
+
+def deploy():
+    set_branch_information()
+    # do_git_stuff()
+    set_ssh_user()
+    provision_server()
+    make_folder_structure()
+    make_virtualenv()
+    create_and_push_archive()
+    unpack()
+    move_confs()
+    upload_settings()
+    update_virtualenv()
+    manage_new_code()
+    switch_symlink()
+    set_permissions()
+    touch_reload_wsgi()
+    restart_nginx()
+    restart_supervisor()
+    clean_tmp_dir()
+
 
 def notify(msg, col=green, wrap='0'):
-    bar = '+-%s-+' % '-'*len(msg)
+    bar = '+-%s-+' % ('-'*len(msg))
     print ''
     print col(bar)
     print _wrap_with(wrap)(col("| %s |" % msg))
@@ -41,17 +49,18 @@ def set_ssh_user():
         env.user = env.remote_user
 
 
-def _set_branch_information():
+def set_branch_information():
     env.last_commit = local('git rev-parse HEAD', capture=True)
     env.branch = local('git symbolic-ref --short HEAD', capture=True)
     env.build_name = '%s%s' % (
-        env.branch, datetime.datetime.now().strftime('%Y-%m-%d-%H-%M'))
+        env.branch, datetime.datetime.now().strftime('%Y-%m-%d-%H-%M-%S'))
     env.build_path = '/tmp/build-%(build_name)s.tar.gz' % env
-    env.code_dir = '%s/%s' % (env.builds_dir, env.build_dir)
+    env.code_dir = '%(project_dir)s/builds/%(build_name)s/' % env
 
 
 @runs_once
 def do_git_stuff():
+    'updates local branch and pushes it.'
     notify('updating branch %(branch)' % env)
     local('git pull origin %(branch)s' % env)
     local('git push origin %(branch)s' % env)
@@ -59,9 +68,7 @@ def do_git_stuff():
 
 def create_and_push_archive():
     notify('Zipping up and putting archive on server')
-    _set_branch_information()
-    local('''git archive --format tar %(branch)s %(web_dir)s
-             | gzip > %(build_path)s ''' % env)
+    local('git archive --format tar %(branch)s %(web_dir)s | gzip > %(build_path)s ' % env)  # noqa
     put(env.build_path, env.build_path)
 
 
@@ -83,12 +90,12 @@ def provision_server():
         'install -y memcached',
         'install -y nginx',
         ]
-
-    for cmd in apt_get_cmds:
-        sudo('apt-get %s' % cmd)
-
-    sudo('apt-get clean &&  apt-get autoclean')
-    sudo('pip install virtualenv')
+    provision = prompt(red('Provision Server? y/N'))
+    if provision.strip() in ['y', 'Y']:
+        for cmd in apt_get_cmds:
+            sudo('apt-get %s' % cmd)
+        sudo('apt-get clean &&  apt-get autoclean')
+        sudo('pip install virtualenv')
 
 
 def make_virtualenv():
@@ -96,8 +103,9 @@ def make_virtualenv():
         '`which virtualenv` --no-site-packages . ',
         'echo "export DJANGO_CONF=\"conf.%(build)s\"" >> /bin/activate' % env,
     ]
+
     if not exists(env.virtualenv):
-        sudo('mkdir -p %(virtualenv_dir)s' % env)
+        sudo('mkdir -p %(virtualenv)s' % env)
         with cd(env.virtualenv):
             [sudo(cmd) for cmd in v_env_commands]
 
@@ -105,35 +113,38 @@ def make_virtualenv():
 def make_folder_structure():
     sudo('mkdir -p %(builds_dir)s' % env)
     cmds = [
-        'mkdir -p builds/%(builds_name)s',
+        'mkdir -p builds/%(build_name)s',
         'mkdir -p data/%(build)s',
         'mkdir -p logs/%(build)s',
         'mkdir -p media/%(build)s',
         'mkdir -p run/%(build)s',
         'chmod -R g+w logs/ media/ run/',
-        '''
-           chown -R %(webserver_user)s:%(webserver_user) -R
-           data/ logs/ media/ run/
-        ''',
+        'chown -R %(webserver_user)s:%(webserver_user)s data/ ',
+        'chown -R %(webserver_user)s:%(webserver_user)s media/',
+        'chown -R %(webserver_user)s:%(webserver_user)s logs/ ',
+        'chown -R %(webserver_user)s:%(webserver_user)s run/',
     ]
     with cd(env.project_dir):
         [sudo(cmd % env) for cmd in cmds]
 
 
 def unpack():
-    # make folder structure
     sudo('mkdir -p %(builds_dir)s' % env)
     with cd(env.builds_dir):
-        sudo('mkdir %(build_name)s')
-        sudo('tar -xzf %(build_path)s -C %(build_name)' % env)
+        sudo('if [ -d "%(build_name)s" ]; then rm -rf "%(build_name)s"; fi' % env)
+        sudo('mkdir %(build_name)s' % env)
+        sudo('tar -xzf %(build_path)s -C %(build_name)s' % env)
+    with cd(env.code_dir):
+        sudo('mv www/* . && rm -rf www')
 
 
-def upload_confs():
+def move_confs():
     'uploading nginx and supervisor confs'
-    confs = [('%(nginx_conf)s' % env, '/etc/nginx/sites-enabled/'),
-             ('%(supervisor_conf)s' % env, 'etc/supervisor/conf.d/'),
-             ('deploy/requirements.txt', '/tmp/'), ]
-    [put(c[0], c[1]) for c in confs]
+    cmds = [
+        'cp deploy/nginx/%(build)s.conf /etc/nginx/sites-enabled/',
+        'cp deploy/supervisord/%(build)s.conf /etc/supervisor/conf.d/']
+    with cd(env.code_dir):
+        [sudo(cmd % env) for cmd in cmds]
 
 
 def touch_reload_wsgi():
@@ -148,10 +159,8 @@ def set_permissions():
         sudo('chmod -R 775 public')
         sudo('chown -R root:%(webserver_user)s public' % env)
 
-# TODO:
-# collect_static, migrate, clean-up (/tmp/), switch symlink
-# reload supervisor
-# reload nginx
+    with cd(env.project_dir):
+        sudo('chown -R www-data:www-data run/ logs/ builds/ media/')
 
 
 def _activate_virtualenv():
@@ -159,10 +168,42 @@ def _activate_virtualenv():
 
 
 def update_virtualenv():
-    sudo(_activate_virtualenv+'pip install -r tmp/requirements.txt' % env)
+    with cd(env.code_dir+'/deploy'):
+        sudo(_activate_virtualenv() + 'pip install -r requirements.txt')
 
 
+def manage_new_code():
+    'Updates database and deals with static files'
+    cmds = [
+        './manage.py syncdb --noinput',
+        './manage.py migrate',
+        './manage.py collectstatic --noinput']
+
+    for cmd in cmds:
+        with cd(env.code_dir):
+            sudo(_activate_virtualenv() + cmd)
 
 
+def upload_settings():
+    put('www/conf/%(build)s.py' % env, '/tmp/')
+    sudo('mv /tmp/%(build)s.py %(code_dir)s/conf/' % env)
 
 
+def restart_nginx():
+    sudo('service nginx restart')
+
+
+def restart_supervisor():
+    # sudo('service supervisor restart ')
+    sudo('supervisorctl restart %(client)s-%(project_code)s-%(build)s' % env)
+
+
+def clean_tmp_dir():
+    sudo('rm -rf /tmp/')
+
+
+def switch_symlink():
+    notify("Switching symlinks")
+    with cd(env.builds_dir):
+        sudo('if [ -h %(build)s ]; then unlink %(build)s; fi' % env)
+        sudo('ln -s %(code_dir)s %(build)s' % env)
